% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_file_nos.R
\name{adj_file_nos}
\alias{adj_file_nos}
\title{Adjust file numbers.}
\usage{
adj_file_nos(target, directory = "munge", action = "up", step = 1)
}
\arguments{
\item{target}{Required. The number in the sequential scripts to begin the
adjustment. The adjustment will affect script with that leading digit and
greater.}

\item{directory}{Defaults to "munge". The directory holding the sequential
scripts.}

\item{action}{Defaults to "up". Whether to adjust file numbers up or down.}

\item{step}{Defaults to 1. The step by which to increment or decrement the
file numbering.}
}
\value{
Renumbers filenames in the specified directory, according to the
specified action. Only affects the target file and above.
}
\description{
This function is used to increment / decrease sequential scripts within the
specified directory, allowing efficient adjustment of script sequence for
additional or removed files.
}
