% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file.R
\name{current_file}
\alias{current_file}
\title{Return the name of the current file.}
\usage{
current_file()
}
\value{
Returns the name of the script currently being executed.
}
\description{
Used to identify the sequential script currently executing. Useful for
logging and debugging.
}
