% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logs.R
\name{log_enable}
\alias{log_enable}
\title{log_enable}
\usage{
log_enable(logfile_loc = "logs/logfile.txt")
}
\arguments{
\item{logfile_loc}{The path to the logfile. Defaults to "logs/logfile.txt".}
}
\value{
Creates logger and file appender.
}
\description{
Assigns the necessary global scope objects for logging with log4r.
}
