% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_file_ops.R
\name{log_file_ops}
\alias{log_file_ops}
\title{log_file_ops}
\usage{
log_file_ops(dir_path = "logs", logfile_nm = "logfile")
}
\arguments{
\item{dir_path}{The name of the folder in which the logfile should be saved.
Creates the folder if required. Defaults to "logs".}

\item{logfile_nm}{Provide a name for the logfile. Do not include suffix.
Defaults to "logfile".}
}
\value{
Creates log directory and log file if required. Calls log_enable()
to assign necessary logging objects in global scope.
}
\description{
Create the necessary file infrastructure to efficiently get up and running
with log4r.
}
\examples{
log_file_ops()

unlink("logs", recursive = TRUE)

}
