% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_report.R
\name{memory_report}
\alias{memory_report}
\title{Perform garbage collection and log allocated memory.}
\usage{
memory_report(logger = my_logger)
}
\arguments{
\item{logger}{the name of the log4r logger object. Defaults to my_logger.}
}
\value{
Performs garbage collection and logs memory size and script name
currently being executed.
}
\description{
Used to log memory allocation at points during sequential script execution.
}
