% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_file_ops.R
\name{sequence_file_ops}
\alias{sequence_file_ops}
\title{sequence_file_ops}
\usage{
sequence_file_ops(target_dir = "munge", n, filetype = "R", force = FALSE)
}
\arguments{
\item{target_dir}{Directory to create files. Defaults to "munge". Creates
the directory if file.exists(target_dir) evaluates to FALSE.}

\item{n}{The number of files to create. Also accepts numerical vector.}

\item{filetype}{The suffix to append the filename. Defaults to ".R".}

\item{force}{Defaults to FALSE. If set to TRUE, sequence_file_ops will
overwrite any pre-existing files that match the write filenames asked for.}
}
\value{
Write a series of sequentially numbered files within a specified
directory. Creates the directory if required.
}
\description{
Quickly create the required number of sequentially labelled files.
}
