% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logs.R
\name{log_enable}
\alias{log_enable}
\title{log_enable}
\usage{
log_enable(logfile_loc = "logs/logfile.txt", pos = 1)
}
\arguments{
\item{logfile_loc}{The path to the logfile. Defaults to "logs/logfile.txt".}

\item{pos}{The environment which to assign pipeline_message. Defaults to 1,
equivalent to the .GlobalEnv.}
}
\value{
Creates logger and file appender.
}
\description{
Assigns the necessary global scope objects for logging with log4r.
}
