% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_report.R
\name{memory_report}
\alias{memory_report}
\title{Perform garbage collection and log allocated memory.}
\usage{
memory_report()
}
\value{
Performs garbage collection then messages memory size and script name
currently being executed.
}
\description{
Used to log memory allocation at points during sequential script execution.
}
\examples{

try(memory_report())

}
