% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_up.R
\name{wrap_up}
\alias{wrap_up}
\title{Wrap up file execution.}
\usage{
wrap_up(start_time, pos = 1)
}
\arguments{
\item{start_time}{A POSIXct object created by executing Sys.time()'.Ensure
this object is assigned prior to calling wrap_up().}

\item{pos}{The environment which to assign pipeline_message. Defaults to 1,
equivalent to the .GlobalEnv.}
}
\value{
Interrupts sequential script execution with an auditory signal. Logs
the elapsed time and outputs the script location.
}
\description{
Used to interrupt sequential script execution while testing or debugging.
messages the elapsed time since execution of \code{start_time}. Assign \code{start_time}
at the start of the first sequential script. Updates the pipeline message.
Outputs an auditory signal and breaks sequential script execution,
identifying the script at which execution was interrupted.
}
