% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_up.R
\name{wrap_up}
\alias{wrap_up}
\title{Wrap up file execution.}
\usage{
wrap_up(
  s_time = start_time,
  pipeline_message,
  logger = my_logger,
  logfile_loc = "logs/logfile.txt"
)
}
\arguments{
\item{s_time}{A POSIXct object created by executing Sys.time()'. Defaults to
start_time. Ensure this object is assigned prior to calling wrap_up().}

\item{pipeline_message}{A character vector object used to convey information
about the current status of the pipeline. Ensure this object is assigned
prior to calling wrap_up().}

\item{logger}{Name of the logger object. Defaults to 'my_logger'.}
}
\value{
Interrupts sequential script execution with an auditory signal. Logs
the elapsed time and outputs the script location.
}
\description{
Used to interrupt sequential script execution while testing or debugging.
Logs the elapsed time since execution of \code{start_time}. Assign \code{start_time} at
the start of the first sequential script. Writes all logged lines to the
logfile. Updates the pipeline message. Outputs an auditory signal and breaks
sequential script execution, identifying the script at which execution was
interrupted.
}
